+++ projects/ngx-mat-facet-search/src/lib/ngx-mat-facet-search.animations.ts
@@ -4,18 +4,23 @@ export const chipAnimation = [
   trigger('chipAnimation', [
     state('in', style({
         transform: 'translateX(0)',
-        opacity: 0
+        opacity: 0,
+        position: 'absolute'
       })
     ),
     transition('void => *', [
       style({
         transform: 'translateX(-100%)',
-        opacity: 1
+        opacity: 1,
+        position: '*'
       }),
-      animate(150)
+      animate('200ms ease-out')
     ]),
     transition('* => void', [
-      animate(150, style({
+      style({
+        position: 'absolute'
+      }),
+      animate('200ms ease-in', style({
         transform: 'translateX(-100%)',
         opacity: 0
       }))