+++ projects/ngx-mat-facet-search/src/lib/ngx-mat-facet-search.module.ts
@@ -23,6 +23,7 @@ import {A11yModule} from '@angular/cdk/a11y';
 import {FacetModalComponent} from './modals/facet-modal/facet-modal.component';
 import {OverlayModule} from '@angular/cdk/overlay';
 import {PortalModule} from '@angular/cdk/portal';
+import { FocusOnShowDirective } from './directives/focus-on-show.directive';
 
 @NgModule({
   imports: [
@@ -56,6 +57,7 @@ import {PortalModule} from '@angular/cdk/portal';
     NgxMatFacetSearchComponent,
     FacetDetailsModalComponent,
     FacetModalComponent,
+    FocusOnShowDirective,
   ],
   exports: [NgxMatFacetSearchComponent],
   entryComponents: [FacetDetailsModalComponent],